requirements.txt
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
dateparser==0.7.6
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
regex==2020.6.8
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tabulate==0.8.7
toml==0.10.1
Twisted==20.3.0
tzlocal==2.1
w3lib==1.22.0
wrapt==1.12.1
yoyo-migrations==7.2.0
zope.interface==5.1.0
